requirements.txt
Django~=3.0.5
djangorestframework~=3.11.0
Sphinx~=3.0.1
sphinx-rtd-theme~=0.4.3
